refactor(frontend): browser compatibility hooks
