+++ frontend/app/components/layouts/public-layout.tsx
@@ -13,8 +13,7 @@ import { PageDetails } from '~/components/page-details';
 import { PageHeaderBrand } from '~/components/page-header-brand';
 import { PageTitle } from '~/components/page-title';
 import { SkipNavigationLinks } from '~/components/skip-navigation-links';
-import { useCurrentLanguage } from '~/hooks';
-import { useBrowserCompatiblityBanner } from '~/hooks/use-browser-compatibility-banner';
+import { useBrowserCompatiblityBanner, useCurrentLanguage } from '~/hooks';
 import { useFeature } from '~/root';
 import * as adobeAnalytics from '~/utils/adobe-analytics.client';
 import { getClientEnv } from '~/utils/env-utils';
