+++ frontend/app/hooks/index.ts
@@ -1,3 +1,5 @@
+export * from './use-browser-compatibility-banner.hook';
+export * from './use-browser-validation.hook';
 export * from './use-current-language.hook';
 export * from './use-enhanced-fetcher.hook';
 export * from './use-fetcher-key.hook';
