+++ frontend/app/hooks/use-browser-validation.hook.ts
